refactor model client/server to transform, add flow_factory tags, split flows by file
