+++ uniflow/constants.py
@@ -2,3 +2,9 @@
 
 ROOT_NAME = "root"
 OUTPUT_NAME = "output"
+
+# Flow Types
+BASIC = "basic"
+EXTRACT = "extract"
+MODEL = "model"
+TRANSFORM = "transform"
