+++ uniflow/extract/config.py
@@ -4,8 +4,16 @@ from dataclasses import dataclass
 
 
 @dataclass
-class Config:
+class ExtractConfig:
     """Extract Config Class."""
+
+    flow_name: str = "ExtractTxtFlow"
+    num_thread: int = 1
+
+
+@dataclass
+class ExtractTxtConfig(ExtractConfig):
+    """Extract Txt Config Class."""
 
     flow_name: str = "ExtractTxtFlow"
     num_thread: int = 1
