+++ uniflow/extract/server.py
@@ -6,8 +6,9 @@ from typing import Any, Dict, List, Mapping, Tuple
 
 from tqdm.auto import tqdm
 
-from uniflow.extract.config import Config
-from uniflow.extract.flow.flow_factory import FlowFactory
+from uniflow.constants import EXTRACT
+from uniflow.extract.config import ExtractConfig
+from uniflow.flow.flow_factory import FlowFactory
 from uniflow.op.op import OpScope
 
 
@@ -22,9 +23,9 @@ class Server:
             num_thread (int, optional): Number of threads to run the flow. Defaults to 1.
         """
         # convert from dict to config for type checking
-        self._config = Config(**config)
+        self._config = ExtractConfig(**config)
 
-        self._flow_cls = FlowFactory.get(self._config.flow_name)
+        self._flow_cls = FlowFactory.get(self._config.flow_name, flow_type=EXTRACT)
         self._num_thread = self._config.num_thread
         self._flow_queue = Queue(self._num_thread)
         for i in range(self._num_thread):
@@ -32,12 +33,12 @@ class Server:
                 self._flow_queue.put(self._flow_cls())
 
     def _run_flow(
-        self, input: Mapping[str, Any], index: int
+        self, input_list: Mapping[str, Any], index: int
     ) -> Tuple[int, Mapping[str, Any]]:
         """Run the flow
 
         Args:
-            input (Mapping[str, Any]): Input to the flow
+            input_list (Mapping[str, Any]): Input to the flow
             index (int): Index of the input
 
         Returns:
@@ -48,31 +49,31 @@ class Server:
         # this is very import to prevent deadlock #
         ###########################################
         try:
-            output = f(input)
+            output = f(input_list)
         except Exception as e:
             output = {"error": str(e)}
         self._flow_queue.put(f)
         return (index, output)
 
     def _run_flow_wrapper(
-        self, input: Mapping[str, Any], i: int
+        self, input_list: Mapping[str, Any], i: int
     ) -> Tuple[int, Mapping[str, Any]]:
         """Wrapper for _run_flow
 
         Args:
-            input (Mapping[str, Any]): Input to the flow
+            input_list (Mapping[str, Any]): Input to the flow
             i (int): Index of the input
 
         Returns:
             Tuple[int, Mapping[str, Any]]: Index of the output, Output from the flow
         """
-        return self._run_flow(input, i)
+        return self._run_flow(input_list, i)
 
-    def run(self, input: List[Mapping[str, Any]]) -> List[Mapping[str, Any]]:
+    def run(self, input_list: List[Mapping[str, Any]]) -> List[Mapping[str, Any]]:
         """Run the flow
 
         Args:
-            input (List[Mapping[str, Any]]): List of inputs to the flow
+            input_list (List[Mapping[str, Any]]): List of inputs to the flow
 
         Returns:
             List[Mapping[str, Any]]: List of outputs from the flow
@@ -80,15 +81,18 @@ class Server:
         with futures.ThreadPoolExecutor(max_workers=self._num_thread) as executor:
             output_futures = {
                 executor.submit(self._run_flow_wrapper, input_data, i): i
-                for i, input_data in enumerate(input)
+                for i, input_data in enumerate(input_list)
             }
-            results = [None] * len(input)
+            results = [None] * len(input_list)
 
-            for future in tqdm(futures.as_completed(output_futures), total=len(input)):
+            for future in tqdm(
+                futures.as_completed(output_futures), total=len(input_list)
+            ):
                 index = output_futures[future]
                 results[index] = future.result()[1]
         return results
 
     def async_run(self):
+        """Run the flow asynchronously"""
         # TODO: Implement async server
         print("Server running async")
